[ADD] Event System, Folder Structure, [EDIT] SDL
Additionally made a proper Event and Eventhandler System
Currently an issue is that Engine.cpp also runs HandleEvents function which will later be within the Input system instead hopefully.

Change the structure to be more manageable for the code and logical.

SDL are now properly included with the Build and allow us to run the .exe within x64 as desired.