[ADD] Comments, InputSystem Expansion
Additionally included more comments to make it more readable and changed some function names to be more clear.
All Keys can now be Debugged to Print out Of Pressed Held down and Released within the EventHandler.

Removed the .json file for now going to include it as Action with the key pressed. So they can be modified within the json file without having the need to do it inside there.

- If Key pressed, and it has an action then do something with the action. -

Input As well has map reading and initialization changes.
CleanUp Added if or when Necessary.

Engine cleanup now assing a nullptr if object is cleaned up to avoid exceptions.